Feature/20230521 embedding files

+++ src/providers/conversationsWebviewProvider/onDidFunctions/onDidConversationDelete.ts
@@ -1,6 +1,17 @@
+import { window } from 'vscode'
 import { IConversation } from '@app/interfaces'
 import { ConversationService } from '@app/services'
 
 export const onDidConversationDelete = (conversation: IConversation): void => {
-  ConversationService.instance.delete(conversation.conversationId)
+  window
+    .showInformationMessage(
+      'Are you sure you want to delete this conversation?',
+      'Yes',
+      'No'
+    )
+    .then((answer) => {
+      if (answer === 'Yes') {
+        ConversationService.instance.delete(conversation.conversationId)
+      }
+    })
 }
